Implements initializer and set_level in Game Class
This commit implements the initializer for the Game Class as well as the
set_level method. It also adds doctests for the set_level method, to
make it easier for users to understand what the method is supposed
to do. The set_level method allows the user to choose which level to
start at when the game first begins.

The commit is essential as the game class is crucial in visualizing the
snake gameplay.