Implements the get_score method in game.py
Implements the get score method in game.py. This is crucial to the 
project as a whole as it allows the the score to be calculated and 
retrieved for the user to see. This update also adds documentation to 
the method so other developers can see how the method works.